+++ lib/vlc_player.dart
@@ -5,6 +5,7 @@
 library;
 
 export 'src/vlc_player_error.dart';
+export 'src/vlc_media_stats.dart';
 export 'src/vlc_media_source.dart';
 export 'src/vlc_media_info.dart';
 export 'src/vlc_player.dart';